Move offer to payment.offer
