+++ eclair-core/src/main/scala/fr/acinq/eclair/payment/offer/OfferManager.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.acinq.eclair.offer
+package fr.acinq.eclair.payment.offer
 
 import akka.actor.typed.scaladsl.Behaviors
 import akka.actor.typed.{ActorRef, Behavior}
